requirements.txt
urllib3==1.23
requests==2.20.0
PyNaCl==1.2.1
gevent==1.3.6
defusedxml==0.5.0
Flask==1.0.2
PySocks==1.6.8
stem==1.6.0
deadsimplekv==0.0.1